Do not assume usize is u32

+++ crates/api-desc/src/lib.rs
@@ -227,7 +227,7 @@ impl Enum {
                 }
             }
             impl From<#name> for isize {
-                fn from(x: #name) -> Self { !(x as u32) as isize }
+                fn from(x: #name) -> Self { !(x as usize) as isize }
             }
         }
     }
